Allow empty IDAT Chunks
ChunkyPNG raises an exception on empty IDAT chunks, however the PNG
specification allows chunks to be empty as long as their length is
correctly set to 0.

Per the W3 PNG specification:

5.3 Chunk layout

Each chunk consists of three or four fields (see figure 5.1). The
meaning of the fields is described in Table 5.1. The chunk data field
may be empty.